Added routes and moved hero cards to new component
