+++ app/public/index.html
@@ -4,8 +4,9 @@
     <meta charset="utf-8" />
 
     <link href="https://fonts.googleapis.com/css?family=Nunito&display=swap" rel="stylesheet">
-
+    <link href="https://fonts.googleapis.com/css?family=Bitter&display=swap" rel="stylesheet">
     <!-- <link rel="icon" href="%PUBLIC_URL%/favicon.ico" /> -->
+    <link rel='icon' href='%PUBLIC_URL%/favicon.ico'/>
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
     <meta
@@ -27,7 +28,7 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
-    <title>React App</title>
+    <title>Dota 2 Stats</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
