+++ app/src/App.css
@@ -1,12 +1,53 @@
 body{
-    /* background-color: #2F363D; */
+    background-color: #2F363D;
 }
 
 h1{
-    font-size: 4rem;
+    font-family: 'Bitter', serif;
+    font-size: 7rem;
     text-align: center;
-    color: #FDE6E3;
+    color: #907F6F;
     margin: 50px 0;
+}
+
+h2{
+    font-size: 2rem;
+    margin: 10px 0;
+    color: white;
+}
+
+h3{
+    font-size: 1.7rem;
+    margin: 10px 0;
+    color: white;
+}
+
+p{
+    font-size: 1.3rem;
+    color: white;
+}
+header{
+    height: 100px;
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+header img{
+    width: 300px;
+    margin-left: 50px;
+}
+
+header nav{
+    margin-right: 50px;
+}
+
+header nav a{
+    color: #907F6F;
+    font-size: 2.5rem;
+    text-decoration: none;
+    margin-left: 20px;
 }
 
 .card-container{
@@ -46,23 +87,6 @@ h1{
     padding: 0 10px;
 }
 
-h2{
-    font-size: 2rem;
-    margin: 10px 0;
-    color: white;
-}
-
-h3{
-    font-size: 1.7rem;
-    margin: 10px 0;
-    color: white;
-}
-
-p{
-    font-size: 1.3rem;
-    color: white;
-}
-
 .role-container{
     display: flex;
     flex-wrap: wrap;
@@ -82,12 +106,16 @@ p{
     padding: 0 5px;
 }
 
-.video{
+video{
     position: fixed;
-    right: 0;
-    bottom: 0;
+    left: 50%;
+    top: 50%;
+    transform: translate(-50%, -50%);
     min-width: 100%;
     min-height: 100%;
     z-index: -1;
+    object-fit: cover;
 }
+
+
