+++ app/src/index.js
@@ -4,6 +4,8 @@ import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 
+import { BrowserRouter as Router } from 'react-router-dom';
+
 import { dataReducer } from './reducers/dataReducer';
 import { createStore, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
@@ -12,9 +14,11 @@ import thunk from 'redux-thunk';
 const store = createStore(dataReducer, applyMiddleware(thunk));
 
 ReactDOM.render(
-<Provider store={store}>
-    <App />
-</Provider>
+<Router>
+    <Provider store={store}>
+        <App />
+    </Provider>
+</Router>
 , document.getElementById('root'));
 
 // If you want your app to work offline and load faster, you can change
